+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Bug fixes
 
 * [#4741](https://github.com/bbatsov/rubocop/issues/4741): Make `Style/SafeNavigation` correctly exclude methods called without dot. ([@drenmi][])
+* [#4740](https://github.com/bbatsov/rubocop/issues/4740): Make `Lint/RescueWithoutErrorClass` aware of modifier form `rescue`. ([@drenmi][])
 
 ## 0.50.0 (2017-09-14)